requirements.txt
gradio
faiss-cpu
numpy
requests
pymongo
Pillow